+++ src/components/Accounts/ChargeFeeFine/ChargeFeeFine.js
@@ -10,7 +10,7 @@ import {
 } from '@folio/stripes/components';
 import {
   FormattedMessage,
-  injectIntl
+  injectIntl,
 } from 'react-intl';
 import SafeHTMLMessage from '@folio/react-intl-safe-html';
 import ChargeForm from './ChargeForm';
@@ -43,6 +43,9 @@ class ChargeFeeFine extends React.Component {
       }),
       account: PropTypes.shape({
         GET: PropTypes.func,
+      }),
+      pay: PropTypes.shape({
+        POST: PropTypes.func.isRequired,
       }),
     }).isRequired,
     stripes: PropTypes.object.isRequired,
@@ -147,29 +150,33 @@ class ChargeFeeFine extends React.Component {
   }
 
   newAction = (action, id, typeAction, amount, comment, balance, transaction) => {
+    const {
+      user,
+      okapi: { currentUser },
+      mutator
+    } = this.props;
     const path = `accounts/${id}`;
     const notify = _.isEmpty(action) ? this.state.notify : this.state.paymentNotify;
 
-    return this.props.mutator.account.GET({ path }).then(record => {
+    return mutator.account.GET({ path }).then(record => {
       const dateAction = _.get(record, ['metadata', 'updatedDate'], moment().format());
 
       const newAction = {
         typeAction,
-        source: `${this.props.okapi.currentUser.lastName}, ${this.props.okapi.currentUser.firstName}`,
-        createdAt: this.props.okapi.currentUser.curServicePoint.id,
+        source: `${currentUser.lastName}, ${currentUser.firstName}`,
+        createdAt: currentUser.curServicePoint.id,
         accountId: id,
         dateAction,
-        userId: this.props.user.id,
+        userId: user.id,
         amountAction: parseFloat(amount || 0).toFixed(2),
         balance: parseFloat(balance || 0).toFixed(2),
         transactionInformation: transaction || '-',
         comments: comment,
         notify,
       };
-      this.props.mutator.feefineactions.POST(Object.assign(action, newAction));
+      mutator.feefineactions.POST(Object.assign(action, newAction));
     });
   }
-
 
   onClickPay(type) {
     this.type = type;
@@ -177,6 +184,7 @@ class ChargeFeeFine extends React.Component {
     this.setState({
       pay: true,
     });
+    this.onClickCharge(type);
     return new Promise((resolve, reject) => {
       this.payResolve = resolve;
       this.payReject = reject;
@@ -202,12 +210,17 @@ class ChargeFeeFine extends React.Component {
 
   onChangeOwner(e) {
     const ownerId = e.target.value;
-    if (_.get(this.props.resources, ['activeRecord', 'shared']) === undefined) {
-      const owners = _.get(this.props.resources, ['owners', 'records'], []);
+    const {
+      resources,
+      mutator,
+    } = this.props;
+
+    if (_.get(resources, ['activeRecord', 'shared']) === undefined) {
+      const owners = _.get(resources, ['owners', 'records'], []);
       const shared = (owners.find(o => o.owner === 'Shared') || {}).id || '0';
-      this.props.mutator.activeRecord.update({ shared });
+      mutator.activeRecord.update({ shared });
     }
-    this.props.mutator.activeRecord.update({ ownerId });
+    mutator.activeRecord.update({ ownerId });
     this.setState({
       ownerId,
     });
@@ -228,18 +241,25 @@ class ChargeFeeFine extends React.Component {
   }
 
   showCalloutMessage(a) {
+    const amount = parseFloat(a.amount).toFixed(2);
+    const paymentName = (a.paymentStatus.name).toLowerCase();
+    const fullName = getFullName(this.props.user);
+    const { feeFineType } = a;
     const message =
       <span>
         <FormattedMessage id="ui-users.charge.messageThe" />
-        {a.feeFineType}
+        {feeFineType}
         <FormattedMessage id="ui-users.charge.messageOf" />
-        <strong>{`${parseFloat(a.amount).toFixed(2)}`}</strong>
+        <strong>{amount}</strong>
         <FormattedMessage id="ui-users.charge.messageSuccessfully" />
-        <strong>{(a.paymentStatus.name).toLowerCase()}</strong>
+        <strong>{paymentName}</strong>
         <FormattedMessage id="ui-users.charge.messageFor" />
-        <strong>{`${getFullName(this.props.user)}`}</strong>
+        <strong>{fullName}</strong>
       </span>;
-    this.callout.sendCallout({ message });
+
+    if (this.callout) {
+      this.callout.sendCallout({ message });
+    }
   }
 
   showConfirmDialog = (values) => {
@@ -263,42 +283,83 @@ class ChargeFeeFine extends React.Component {
   }
 
   onConfirm = () => {
-    const { values } = this.state;
-    const { intl: { formatMessage } } = this.props;
+    const { values, pay, notify } = this.state;
+    const {
+      intl: { formatMessage },
+      okapi: {
+        currentUser: {
+          curServicePoint: { id: servicePointId }
+        },
+      },
+      user,
+      mutator,
+    } = this.props;
     const tagStaff = formatMessage({ id: 'ui-users.accounts.actions.tag.staff' });
     const tagPatron = formatMessage({ id: 'ui-users.accounts.actions.tag.patron' });
     let comment = '';
+
     if (values.comment) {
       comment = `${tagStaff} : ${values.comment}`;
     }
     if (values.patronInfo && values.notify) {
       comment = `${comment} \n ${tagPatron} : ${values.patronInfo}`;
     }
-    this.onClickCharge(this.type).then(() => {
-      this.type.remaining = parseFloat(this.type.amount - values.amount).toFixed(2);
-      let paymentStatus = _.capitalize(formatMessage({ id: 'ui-users.accounts.actions.warning.paymentAction' }));
-      if (this.type.remaining === '0.00') {
-        paymentStatus = `${paymentStatus} ${formatMessage({ id: 'ui-users.accounts.status.fully' })}`;
-        this.type.status.name = 'Closed';
-      } else {
-        paymentStatus = `${paymentStatus} ${formatMessage({ id: 'ui-users.accounts.status.partially' })}`;
-      }
-      this.type.paymentStatus.name = paymentStatus;
-      this.props.mutator.activeRecord.update({ id: this.type.id });
 
-      return this.props.mutator.accounts.PUT(_.omit(this.type, ['comments', 'patronInfo', 'notify']));
-    })
-      .then(() => this.newAction({ paymentMethod: values.method }, this.type.id,
-        this.type.paymentStatus.name, values.amount,
-        comment, this.type.remaining,
-        values.transaction, this.type.feeFineOwner))
-      .then(() => {
-        this.hideConfirmDialog();
-        this.onClosePayModal();
-        this.showCalloutMessage(this.type);
-        this.payResolve();
-      })
-      .then(() => { this.props.history.goBack(); });
+    this.type.remaining = parseFloat(this.type.amount - values.amount).toFixed(2);
+    let paymentStatus = _.capitalize(formatMessage({ id: 'ui-users.accounts.actions.warning.paymentAction' }));
+
+    if (this.type.remaining === '0.00') {
+      paymentStatus = `${paymentStatus} ${formatMessage({ id: 'ui-users.accounts.status.fully' })}`;
+      this.type.status.name = 'Closed';
+    } else {
+      paymentStatus = `${paymentStatus} ${formatMessage({ id: 'ui-users.accounts.status.partially' })}`;
+    }
+
+    this.type.paymentStatus.name = paymentStatus;
+    mutator.activeRecord.update({ id: this.type.id });
+
+    if (pay) {
+      const payBody = {
+        amount: values.amount,
+        notifyPatron: notify,
+        servicePointId,
+        userName: getFullName(user),
+        paymentMethod: values.method,
+        comments: comment,
+        transactionInfo: values.transaction
+      };
+
+      return mutator.pay.POST(payBody)
+        .then(this.updateAccountAfterPay(values, comment))
+        .then(this.showCalloutMessage(this.type));
+    } else {
+      return this.updateAccountAfterPay(values, comment);
+    }
+  }
+
+  updateAccountAfterPay = (values, comment) => {
+    const type = this.type;
+    return this.props.mutator.accounts.PUT(_.omit(type, ['comments', 'patronInfo', 'notify']))
+      .then(() => this.newAction(
+        { paymentMethod: values.method },
+        type.id,
+        type.paymentStatus.name,
+        values.amount,
+        comment,
+        type.remaining,
+        values.transaction,
+        type.feeFineOwner
+      ))
+      .then(() => this.hideConfirmDialog())
+      .then(() => this.onClosePayModal())
+      .then(() => this.payResolve())
+      .catch((error) => {
+        this.payReject();
+        this.callout.current.sendCallout({
+          type: 'error',
+          message: error
+        });
+      });
   }
 
   renderConfirmMessage = () => {
@@ -319,17 +380,17 @@ class ChargeFeeFine extends React.Component {
   }
 
   onSubmitCharge = (data) => {
-    const {
-      history
-    } = this.props;
+    const { history } = this.props;
+
     if (data.pay) {
       delete data.pay;
       this.type.remaining = data.amount;
-      this.onClickPay(data);
+      this.onClickPay(data)
+        .then(() => history.goBack());
     } else {
       delete data.pay;
       this.onClickCharge(data)
-        .then(() => { history.goBack(); });
+        .then(() => history.goBack());
     }
   }
 
@@ -359,8 +420,8 @@ class ChargeFeeFine extends React.Component {
     const feefines = (this.state.ownerId !== '0') ? (resources.feefines || {}).records || [] : [];
     const payments = _.get(resources, ['payments', 'records'], []).filter(p => p.ownerId === this.state.ownerId);
     const accounts = _.get(resources, ['accounts', 'records'], []);
-    const settings = _.get(this.props.resources, ['commentRequired', 'records', 0], {});
-    const barcode = _.get(this.props.resources, 'activeRecord.barcode');
+    const settings = _.get(resources, ['commentRequired', 'records', 0], {});
+    const barcode = _.get(resources, 'activeRecord.barcode');
 
     const defaultServicePointId = _.get(resources, ['curUserServicePoint', 'records', 0, 'defaultServicePointId'], '-');
     const servicePointsIds = _.get(resources, ['curUserServicePoint', 'records', 0, 'servicePointsIds'], []);
@@ -398,7 +459,9 @@ class ChargeFeeFine extends React.Component {
     return (
       <div>
         <ChargeForm
+          form="feeFineChargeForm"
           onClickPay={this.onClickPay}
+          initialValues={initialValues}
           defaultServicePointId={defaultServicePointId}
           servicePointsIds={servicePointsIds}
           onSubmit={this.onSubmitCharge}
@@ -440,6 +503,8 @@ class ChargeFeeFine extends React.Component {
           onSubmit={(values) => { this.showConfirmDialog(values); }}
           owners={owners}
           feefines={feefines}
+          okapi={this.props.okapi}
+          checkAmount="check-pay"
         />
         <Callout ref={(ref) => { this.callout = ref; }} />
         <ConfirmationModal