+++ src/components/util/util.js
@@ -15,11 +15,6 @@ export function getFullName(user) {
   return `${lastName}${firstName ? ', ' : ' '}${firstName}${middleName ? ' ' : ''}${middleName}`;
 }
 
-export function eachPromise(arr, fn) {
-  if (!Array.isArray(arr)) return Promise.reject(new Error('Array not found'));
-  return arr.reduce((prev, cur) => (prev.then(() => fn(cur))), Promise.resolve());
-}
-
 export function validate(item, index, items, field, label) {
   const error = {};
   for (let i = 0; i < items.length; i++) {
@@ -82,10 +77,10 @@ export function getOpenRequestStatusesFilterString() {
   return getFilterStatusesString(openStatusesArr);
 }
 
-export function getOpenRequestsPath(barcode) {
+export function getOpenRequestsPath(itemId) {
   const filterString = getOpenRequestStatusesFilterString();
 
-  return `/requests?filters=${filterString}&query=${barcode}&sort=Request Date`;
+  return `/requests?filters=${filterString}&query=${itemId}&sort=Request Date`;
 }
 
 export function getChargeFineToLoanPath(userId, loanId) {
@@ -113,6 +108,13 @@ export function calculateSortParams({
   return sortParams;
 }
 
+// Return true if every item in loans has the status itemStatus
 export function hasEveryLoanItemStatus(loans, itemStatus) {
   return _.every(Object.values(loans), loan => loan?.item?.status?.name === itemStatus);
 }
+
+// Return true if every item in loans has one of the statuses in the itemStatuses array
+export function hasAnyLoanItemStatus(loans, itemStatuses) {
+  return _.every(Object.values(loans), loan => itemStatuses.includes(loan?.item?.status?.name));
+}
+