+++ src/settings/patronBlocks/Conditions/Conditions.js
@@ -8,6 +8,7 @@ import SafeHTMLMessage from '@folio/react-intl-safe-html';
 import ConditionsForm from './ConditionsForm';
 import css from '../patronBlocks.css';
 
+
 class Conditions extends Component {
   static manifest = Object.freeze({
     patronBlockCondition: {
@@ -23,6 +24,9 @@ class Conditions extends Component {
   });
 
   static propTypes = {
+    resources: PropTypes.shape({
+      patronBlockCondition: PropTypes.object
+    }),
     id: PropTypes.string.isRequired,
     mutator: PropTypes.shape({
       patronBlockCondition: PropTypes.shape({