+++ src/views/LoanDetails/LoanProxyDetails.js
@@ -1,10 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import Link from 'react-router-dom/Link';
+import { Link } from 'react-router-dom';
 import { KeyValue } from '@folio/stripes/components';
 import { stripesConnect } from '@folio/stripes/core';
 import { getFullName } from '../../components/util';
+
+
 
 class LoanProxyDetails extends React.Component {
   static manifest = Object.freeze({