+++ test/bigtest/tests/charge-fee-fine-test.js
@@ -71,7 +71,6 @@ describe('Charge fee/fine', () => {
           describe('cancel the charge', () => {
             beforeEach(async () => {
               await chargeForm.clickCancel();
-              await chargeForm.clickConfirmCancel();
             });
 
             it('navigate to previous page', function () {